[#4737] replace annotation key from common
