--- bootstrap-core/src/main/java/com/navercorp/pinpoint/bootstrap/plugin/test/ExpectedMongoJson.java
@@ -1,50 +1,0 @@
-/*
- * Copyright 2018 NAVER Corp.
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-package com.navercorp.pinpoint.bootstrap.plugin.test;
-
-import com.navercorp.pinpoint.common.trace.AnnotationKey;
-
-/**
- * @author Roy Kim
- */
-public class ExpectedMongoJson extends ExpectedAnnotation {
-    private final String bindValue;
-
-    public ExpectedMongoJson(String query, String bindValue) {
-        super(AnnotationKey.MONGO_JSON.getName(), query);
-        this.bindValue = bindValue;
-    }
-
-    public String getQuery() {
-        return (String) getValue();
-    }
-
-    public String getBindValue() {
-        return bindValue;
-    }
-
-    @Override
-    public String toString() {
-        StringBuilder builder = new StringBuilder();
-        builder.append(getKeyName());
-        builder.append("=[query=");
-        builder.append(getQuery());
-        builder.append(", bindValue=");
-        builder.append(bindValue);
-        builder.append("]");
-
-        return builder.toString();
-    }
-}
