+++ plugins/mongodb/src/main/java/com/navercorp/pinpoint/plugin/mongo/MongoTypeProvider.java
@@ -26,7 +26,13 @@ public class MongoTypeProvider implements TraceMetadataProvider {
 
     @Override
     public void setup(TraceMetadataSetupContext context) {
-        context.addServiceType(MongoConstants.MONGO, AnnotationKeyMatchers.exact(AnnotationKey.ARGS0));
+        context.addServiceType(MongoConstants.MONGODB, AnnotationKeyMatchers.exact(AnnotationKey.ARGS0));
         context.addServiceType(MongoConstants.MONGO_EXECUTE_QUERY, AnnotationKeyMatchers.exact(AnnotationKey.ARGS0));
+
+        context.addAnnotationKey(MongoConstants.MONGO_COLLECTION_INFO);
+        context.addAnnotationKey(MongoConstants.MONGO_COLLECTION_OPTION);
+        context.addAnnotationKey(MongoConstants.MONGO_JSON);
+        context.addAnnotationKey(MongoConstants.MONGO_JSON_BINDVALUE);
+
     }
 }
