+++ web/src/main/java/com/navercorp/pinpoint/web/util/MongoJsonParser.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.common.util;
+package com.navercorp.pinpoint.web.util;
 
 import java.util.List;
