+++ web/src/test/java/com/navercorp/pinpoint/web/util/OutputParameterMongoJsonParserTest.java
@@ -14,7 +14,7 @@
  * limitations under the License.
  */
 
-package com.navercorp.pinpoint.common.util;
+package com.navercorp.pinpoint.web.util;
 
 import org.junit.Assert;
 import org.junit.Test;
